requirements.txt

flask
line-bot-sdk
googlemaps
gunicorn
